Create BetterCord
This is not going to used for personal gain until I have practiced enough and can make it different from discord. And then maybe I will think about future plans. Rn this is solely for practice and use of my code is not permitted for your use. Please do not copy.